feat(drop FFI-nonspecific ops from sector-builder-ffi crate)
