--- sector-builder-ffi/src/error.rs
@@ -1,1 +1,0 @@
-pub type Result<T> = std::result::Result<T, failure::Error>;
