+++ sector-builder-ffi/src/lib.rs
@@ -2,16 +2,13 @@
 
 // These need to be here because of cbindgen: https://github.com/eqrion/cbindgen/issues/292
 extern crate filecoin_proofs;
+extern crate filecoin_proofs_ffi;
 extern crate sector_builder;
 
 #[macro_use]
 extern crate lazy_static;
-#[macro_use]
-extern crate failure;
 
-mod error;
 mod responses;
 mod singletons;
 
-pub mod builder;
-pub mod stateless;
+pub mod api;
